fix: address CodeRabbit minor findings on commit 4c96709
Plumbing and edge-case fixes flagged in the latest CodeRabbit pass.

Edge-case guards:
- ipw: raise ValueError when the treatment column has only one class,
  before LogisticRegression.fit produces an opaque sklearn error.
- timestamps: invalid_indices preserves non-numeric pandas indices
  (datetime, string, MultiIndex) by falling back to str() instead of
  crashing on int().
- placebo: cast the permuted treatment to int before the ==1 mask so
  bool/object dtypes work alongside int/float.

Routing fix:
- router: aggregate_time_series with no pre-period and no comparison
  group now returns not_identifiable instead of falling through to
  doubly_robust_estimation. There is no counterfactual to fit ITS
  against in that configuration.

Robustness:
- synthetic_control: pre_period_mspe diagnostic now compares MSPE to the
  pre-period variance of the treated unit, so the threshold is
  scale-invariant (binary outcomes and revenue dollars get the same
  signal).
- pipeline: skip computing baseline_outcome_rate for non-binary outcomes
  (sensitivity_check ignores it anyway), and remove a dead try/raise
  wrapper around the placebo estimator callback.

Tests: new coverage for IPW single-class guard and the
aggregate_time_series + no-pre-period not_identifiable path. 126 tests
pass; lint clean.

+++ src/causal_debugger/refutation/placebo.py
@@ -35,7 +35,8 @@ def placebo_treatment_test(
             ),
             "delta_vs_main_estimate": None,
         }
-    shuffled = rng.permutation(df[treatment].values)
+    # Cast to int so masks work for bool/object dtypes too.
+    shuffled = rng.permutation(df[treatment].astype(int).values)
     if estimator is not None:
         permuted = df.assign(**{treatment: shuffled})
         try:
